continue gutting lfs/scanner.go
